+++ src/serializer.rs
@@ -1,3 +1,7 @@
+use alloc::borrow::ToOwned;
+use alloc::format;
+use alloc::string::{String, ToString};
+
 use crate::error::Error;
 use crate::value::{Table, Value};